refactoring of sorting module
