+++ src/app.jsx
@@ -1,46 +1,8 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import Users from "./components/users";
-import api from "./api";
 
 const App = () => {
-    // const [users, setUsers] = useState(api.users.fetchAll());
-    const [users, setUsers] = useState();
-    useEffect(() => {
-        api.users.fetchAll().then((data) => setUsers(data));
-    }, []);
-
-    const handleDelete = (userId) => {
-        setUsers((prevState) =>
-            prevState.filter((user) => user._id !== userId)
-        );
-    };
-
-    const handleToggleBookMark = (id) => {
-        // const newUsers = [...users];
-        // const index = newUsers.findIndex((user) => user._id === id);
-        // newUsers[index].bookmark = !newUsers[index].bookmark;
-        // setUsers(newUsers);
-        setUsers(
-            users.map((user) => {
-                if (user._id === id) {
-                    return { ...user, bookmark: !user.bookmark };
-                }
-                return user;
-            })
-        );
-    };
-
-    return (
-        <>
-            {users && users.length !== 0 && (
-                <Users
-                    users={users}
-                    onDelete={handleDelete}
-                    onToggleBookMark={handleToggleBookMark}
-                />
-            )}
-        </>
-    );
+    return <Users/>
 };
 
 export default App;
